+++ ValheimEnforcer/ValheimEnforcer.cs
@@ -10,6 +10,7 @@ using System.Reflection;
 using UnityEngine;
 using ValheimEnforcer.modules;
 using ValheimEnforcer.modules.character;
+using ValheimEnforcer.modules.cheatmonitor;
 using ValheimEnforcer.modules.commands;
 using ValheimEnforcer.modules.compat;
 
@@ -23,7 +24,7 @@ namespace ValheimEnforcer
     {
         public const string PluginGUID = "MidnightsFX.ValheimEnforcer";
         public const string PluginName = "ValheimEnforcer";
-        public const string PluginVersion = "0.8.2";
+        public const string PluginVersion = "0.9.0";
 
         internal static ManualLogSource Log;
         internal ValConfig cfg;
@@ -39,7 +40,7 @@ namespace ValheimEnforcer
             // Just needs to run AFTER all mods are loaded
             // For client
             PrefabManager.OnPrefabsRegistered += ModManager.SetModsActive;
-            PrefabManager.OnPrefabsRegistered += InternalDataStore.InstanciateOrLinkMetadataRegistry;
+            ZoneManager.OnLocationsRegistered += InternalDataStore.InstanciateOrLinkMetadataRegistry;
             // For server
             PrefabManager.OnVanillaPrefabsAvailable += ModManager.SetModsActive;
             GUIManager.OnCustomGUIAvailable += ModManager.AddErrorMessageDetailsForMenu;